test(react-grab): lazy-load + gate the e2e dials demo behind ?dials
Keeps the always-on demo out of the default page so it can't intercept the
selection prompt other specs click, and out of the static import graph so the
perf baseline build doesn't depend on PR-only react-grab/react exports.
